planetary-computer-mcp-py/pyproject.toml
[project]
name = "planetary-computer-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "planetary-computer>=1.0.0",
    "pystac-client>=0.8.0",
    "odc-stac>=0.3.0",
    "xarray>=2024.1.0",
    "rioxarray>=0.15.0",
    "duckdb>=1.0.0",
    "zarr>=2.18.0",
    "matplotlib>=3.8.0",
    "geopy>=2.4.0",
    "geopandas>=0.14.0",
    "pillow>=10.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
pythonpath = ["src"]
